iio: adc: ad4134: Register first and second ADC devices differently
When two ADCs are used together, the first ADC provides the channels
physically present in the second device while the second ADC IIO device
is used only for configuration and should have no channels.
Register the first and second AD4134/AD7134 devices differently to
provide the expected interface and functionality.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>